cleanup(nextjs): only re-export values in index.ts
